buy and sell use case diagram
The buy shares use case includes requesting a list of current stock market shares, which includes connecting to a server. Buy shares additionally has a include of list profit so that the user can tell which market shares they can afford.  The use case also as an extend of not enough funds to buy the share, for when the user is trying to buy a share they can’t afford. 
To sell shares the server first lists currently owned shares which requires it to connect to the server. After the current list of shares are listed the user can choose which shares they want to sell.